requirements.txt
# Web Framework
flask==3.0.0

# Automation
playwright==1.40.0

# Database
sqlalchemy==2.0.23
flask-sqlalchemy==3.1.1

# HTTP Requests
requests==2.31.0
httpx==0.25.2

# Environment
python-dotenv==1.0.0

# Scraping APIs
apify-client==1.6.3

# Proxy
requests[socks]==2.31.0

# Utilities
python-dateutil==2.8.2
